requirements.txt
# CUDA-enabled PyTorch wheels
--extra-index-url https://download.pytorch.org/whl/cu128

# Core ML / GPU runtime
torch==2.11.0+cu128
torchvision==0.26.0+cu128
torchaudio==2.11.0+cu128
transformers==5.5.3
sentence-transformers==5.4.0
peft==0.18.1
bitsandbytes==0.49.2
accelerate==1.13.0
trl==1.0.0
datasets==4.8.4

# RAG and orchestration
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.2.28
langchain-openai==1.1.12
faiss-cpu==1.13.2
ragas==0.4.3

# Document processing
pymupdf==1.27.2.2
pdfplumber==0.11.9

# Tracking and production services
mlflow==3.11.1
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
python-multipart==0.0.24

# Utilities
python-dotenv==1.2.2
loguru==0.7.3
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
pandas==2.3.3
numpy==2.4.4

# Testing
pytest==9.0.3
pytest-asyncio==1.3.0
